refactor: create folder for rooms page and divide components to separate files which belong to rooms page
