--- frontend/src/components/AddRoomForm/AddRoomForm.scss
@@ -1,7 +1,0 @@
-@import '../../App';
-@import '../../share/Card/Card.scss';
-
-.room-form {
-	margin: 10px;
-	max-height: 420px;
-}
