+++ frontend/src/components/RoomsPage/AddRoomForm/AddRoomForm.scss
@@ -1,0 +1,7 @@
+@import '../../../App';
+@import '../../../share/Card/Card.scss';
+
+.room-form {
+    margin: 10px;
+    max-height: 420px;
+}
