+++ frontend/src/components/RoomsPage/RoomsList/RoomsList.scss
