+++ frontend/src/router/AdminRouter.js
@@ -21,7 +21,7 @@ import TeacherList from '../containers/TeachersList/TeachersList';
 import BusyRooms from '../containers/BusyRooms/BusyRooms';
 import GroupList from '../containers/GroupPage/GroupPage';
 import ClassSchedule from '../containers/ClassSchedule/ClassSchedule';
-import RoomList from '../containers/RoomList/RoomList';
+import RoomPage from '../components/RoomsPage/RoomsPage';
 import SubjectPage from '../containers/SubjectPage/SubjectPage';
 import SemesterPage from '../containers/SemesterPage/SemesterPage';
 import MergeRolePage from '../containers/MergeRolePage/MergeRolePage';
@@ -40,7 +40,7 @@ export default function Routers() {
                 <AdminRoute path={TEACHER_LIST_LINK} component={TeacherList} />
                 <AdminRoute path={GROUP_LIST_LINK} component={GroupList} />
                 <AdminRoute path={CLASS_SCHEDULE_TITLE_LINK} component={ClassSchedule} />
-                <AdminRoute path={ROOM_LIST_LINK} component={RoomList} />
+                <AdminRoute path={ROOM_LIST_LINK} component={RoomPage} />
                 <AdminRoute path={SUBJECT_PAGE_LINK} component={SubjectPage} />
                 <AdminRoute path={BUST_ROOMS_LINK} component={BusyRooms} />
                 <AdminRoute path={SEMESTER_PAGE_LINK} component={SemesterPage} />
